Removed unnecessary code duplication in the Maven builders
